Moclobemide-Market.md
Moclobemide is a reversible inhibitor of monoamine oxidase type A (RIMA) antidepressant. It is utilized in the treatment of depression and social anxiety disorders among adults. Moclobemide offers a safer profile compared to irreversible monoamine oxidase inhibitors (MAOIs) due to its reversible nature, making it a preferred choice in clinical practice.