Freeze-dried-Cell-Culture-Medium-Market.md
Freeze-dried Cell Culture Medium refers to a type of culture medium used in cell culture processes that has been freeze-dried to ensure long-term stability and ease of use. This medium contains all the necessary nutrients, growth factors, and supplements required for the growth and maintenance of cell cultures in a laboratory setting. Freeze-dried Cell Culture Medium eliminates the need for regular preparation and can be easily reconstituted by adding water.